Minor cleanup
Also broke up the `cpnam install` into several smaller commands, just to make it easier to repeate the docker build - if any of these modules fail to install, docker will still have the cache of the prior `cpanm install` that succeeded.

Yes, the most stable solution is to put each module on its on RUN command, that can be done later if necessary.

Also, probably way more modules than needed. This list of modules was derived by looking at the output of `cpan -l` on the production server, and then determining what was not yet in the base image. It's possible that not all of these are necessary for the Reactome CGI scripts.