Added a "wait-for" script
I found that my WordPress container was starting too fast (or maybe the database was starting too slow?) so I needed to add a script to wait until the database container completed starting up and was available for use.
Testing the database container is done with nc ("netcat") which needed to be installed since it isn't in WordPress by default.