Modularized the build process of apps
Modified maven_builds.sh: The build process of each separable module is contained in a function. These functions are called by the loop which is present in the last. There is an 'app_list' which contains list of all the functions and their desired state. If they are in ready state, their build process will be skipped.