Created soft link to internal logs of mysql
This use case will run in selected few users. When chown fails to change permissions. We are providing soft links to the internal directory of container. Hard link cannot be given, since they are allowed only for files. And when we create files from here, it caused problems for mysql.