Be more specific about ignoring neo4j
neo4j.conf should not be ignored - it has important stuff in it, but it will be updated everytime Neo4j runs.
TODO: Maybe it should be moved to some sort of staging directory, and the deploy script can copy it to where it needs to be and THAT instance of the file will be ignored.