A few small changes...
Add a small SQL script to updated the home and siteurl variables in WordPress.
This allows us to download it on demand, and also to remove it from github.

Also, chown the neo4j files before cleanup - it is possible that some of the files that deploy.sh tries to clean won't be owned by $USER so they can't be cleaned, and that causes the script to fail.