Continue work on Curator-specific Reactome
The correct Curator-verion of PathwayBrowser is a private repo so you
must have credentials to check it out. Clone it in to
java-application-builder before attempting to build anything.

Also added a custom header.php for Curator use - it uses the old Perl
search functionality instad of the new solr/neo4j search.