Fix logging issues with diagram-converter
It used to log EVERYTHING to the console in DEBUG mode. Apache HTTP, Spring, Neo4j were all dumping tonnes of DEBUG messages to the console...
As it turns out, the pom file was excluding logback.xml. Experimenting with _adding_ to logback.xml (as well as removing the exclusion) and log4j.properties to make logging output more usable.